+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+- Duplicate monomers in a complex composition now have their counts summed instead of raising an error.
+
 ## [0.1.0] - 2026-03-10
 
 - Initial release